Refactor utility container stuff to avoid galaxy -> tool shed imports.
