+++ lib/galaxy/tool_shed/util/utility_container_manager.py
